+++ src/flockwave/server/ext/http_server/extension.py
@@ -301,6 +301,14 @@ async def run(app: SkybrushServer, configuration: dict[str, Any], logger: Logger
     config.keyfile = configuration.get("keyfile") or None
     config.use_reloader = False
 
+    config.websocket_max_message_size = int(
+        configuration.get("websocket_max_message_size", 100 * 1024 * 1024)
+    )
+    logger.debug(
+        f"WebSocket maximum message size is set to "
+        f"{config.websocket_max_message_size // 1024 // 1024} MB"
+    )
+
     secure = bool(config.ssl_enabled)
 
     # Test quickly whether we can bind to the given host and port; if we cannot,
@@ -403,13 +411,26 @@ schema = {
         "certfile": {
             "type": "string",
             "title": "Certificate file",
-            "description": "Full path to the certificate file that the server should use for HTTPS connections",
+            "description": (
+                "Full path to the certificate file that the server should use for "
+                "HTTPS connections"
+            ),
             "required": False,
         },
         "keyfile": {
             "type": "string",
             "title": "Private key file",
             "description": "Full path to the private key file corresopnding to the certificate",
+            "required": False,
+        },
+        "websocket_max_message_size": {
+            "type": "integer",
+            "title": "Max WebSocket message size",
+            "description": (
+                "Maximum size of a single WebSocket message in bytes. Used only for "
+                "clients that connect to the server via WebSocket."
+            ),
+            "default": 104857600,
             "required": False,
         },
     }